Refactored commons and classes out to header files.
